Put content hash in path instead of query string (reverting previous version)
After some reading (see below), it turns out that putting the content hash
in the path is more robust. This does require some additional work for the
app to make work, but that's most likely worth it.

We could make this behavior configurable in some future version.

Reference: https://www.alainschlesser.com/bust-cache-content-hash/